\name{CH1903}
\alias{CH1903}
\docType{data}
\title{
  \code{CH1903}
}
\description{
  Swiss square projection
}
\usage{data(CH1903)}
\format{
  The format is:
Formal class 'CRS' [package "sp"] with 1 slots
  ..@ projargs: chr " +init=world:CH1903 +proj=somerc +lat_0=46d57'8.660\"N +lon_0=7d26'22.500\"E +ellps=bessel +x_0=600000 +y_0=200000 +k_0=1. +no_"| __truncated__
}
\keyword{datasets}
\keyword{spatial}
