\name{GPS}
\alias{GPS}
\docType{data}
\title{
  \code{GPS}
}
\description{
  describe me
}
\usage{data(GPS)}
\format{
  The format is:
  Formal class 'CRS' [package "sp"] with 1 slots
  ..@ projargs: chr " +proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
}
\examples{
data(GPS)
## maybe str(GPS) ; plot(GPS) ...
}
\keyword{datasets}
