\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,MultiKeyHash,ANY,ANY,ANY-method}
\alias{[[-methods}
\alias{[[,MultiKeyHash-method}
\alias{print-methods}
\alias{print,ResampleExactSpawMLOutput-method}
\alias{print,SpawMLOutput-method}
\alias{VarCorr-methods}
\alias{VarCorr,SpawMLOutput-method}
\alias{AIC-methods}
\alias{AIC,SpawMLOutput-method}
\alias{BIC-methods}
\alias{BIC,SpawMLOutput-method}
\alias{fixef-methods}
\alias{fixef,SpawMLOutput-method}
\alias{ranef-methods}
\alias{ranef,SpawMLOutput-method}
\alias{show-methods}
\alias{show,ResampleExactSpawMLOutput-method}
\alias{show,SpawMLOutput-method}
\alias{summary-methods}
\alias{summary,SpawMLOutput-method}
\alias{[<--methods}
\alias{[<-,MultiKeyHash,ANY,ANY,ANY-method}
\alias{[[<--methods}
\alias{[[<-,MultiKeyHash,ANY,ANY-method}
\alias{[[-methods}
\alias{[[,MultiKeyHash,ANY,ANY-method}

\title{helper functions}
\description{self-explanatory helper functions}
\section{Methods}{
\describe{

\item{\code{signature(x = "MultiKeyHash", i = "ANY", j = "ANY", drop = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
